+++ src/heaps/heap23.h
@@ -44,7 +44,7 @@ class Heap23Node {
         Heap23Node *child;
         Heap23Node *left, *right;
         unsigned int dim;
-        float key;
+        double key;
         unsigned int item;
 };
 
@@ -72,9 +72,9 @@ class Heap23 : public Heap {
         Heap23(unsigned int n);
         ~Heap23();
 
-        void insert(unsigned int item, float k);
+        void insert(unsigned int item, double k);
         unsigned int deleteMin();
-        void decreaseKey(unsigned int item, float newValue);
+        void decreaseKey(unsigned int item, double newValue);
         unsigned int nItems() const { return itemCount; }
 
         long int nComps() const { return compCount; }
