+++ src/heaps/radixheap.h
@@ -21,8 +21,8 @@ class RadixHeap: public Heap {
         ~RadixHeap();
 
         unsigned int deleteMin();
-        void insert(unsigned int item, float k);
-        void decreaseKey(unsigned int item, float newValue);
+        void insert(unsigned int item, double k);
+        void decreaseKey(unsigned int item, double newValue);
         unsigned int nItems() const { return itemCount; }
 
         long int nComps() const { return compCount; }
